platform: nordic_nrf: Create a nrf54l folder
Create a nrf54l folder which can be used for other
devices in the series as well.

This is preparation work for the addition of the
nrf54l10.

This also updates the the source and header files
which use ifdefs with nrf54l15 specific defines with
the macro NRF54L_SERIES.

Change-Id: I1b2a0876fefef3fb460cf8995b9b802aa089b7aa
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>